% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_to_scopus.R
\name{MergeToScopus}
\alias{MergeToScopus}
\title{Merge a Web of Science bib file and a Scopus bib file}
\usage{
MergeToScopus(Scopus, Wos)
}
\arguments{
\item{Scopus}{A path to a .bib file from a Scopus search.}

\item{Wos}{A path to a .bib file from a Web of Science search.}
}
\value{
A data frame with the results in Scopus format
}
\description{
Merge a Web of Science bib file and a Scopus bib file
}
\examples{

Test <- MergeToScopus(Scopus = system.file("extdata/scopus16aug.bib", package="BiblioUtils"),
                      Wos = system.file("extdata/WOS11aug.bib", package="BiblioUtils"))

}
